Cleaning up warnings.
And some slightly less fancy 2.10/2.11 struts.

+++ text/src/main/scala/basis/text/UTF.scala
@@ -34,10 +34,10 @@ trait UTF extends Any with Equals with Family[UTF] with Seq[Int] {
         case c if (c >= 0x0000 && c <= 0x001F) ||
                   (c >= 0x007F && c <= 0x009F) =>
           builder.append('\\'); builder.append('u')
-          builder.append(hexToChar(c >>> 12 & 0xF))
-          builder.append(hexToChar(c >>>  8 & 0xF))
-          builder.append(hexToChar(c >>>  4 & 0xF))
-          builder.append(hexToChar(c        & 0xF))
+          builder.append(hexToChar(c.toInt >>> 12 & 0xF))
+          builder.append(hexToChar(c.toInt >>>  8 & 0xF))
+          builder.append(hexToChar(c.toInt >>>  4 & 0xF))
+          builder.append(hexToChar(c.toInt        & 0xF))
         case c => builder.append(c)
       }
       cs.step()
